finally fixing bug where more page would not load
the issue was related to signalr i believe. when you left the original
web browser, the original browser is kept open with signal r. new brwosers
try to do stuff, but nothing happens. the headers do seem to go through,
which is odd, but otherwise it doesn't work.

so the fix is stopping the web browser when it disappears, this will also
stop the updates when you go back however.